LLM-TRAINING/AI_AGENTS_POCS/TRAVEL-AGENT-01/travel-backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv

# LangChain stack
langchain
langgraph
langchain-openai
langchain-community
langchain-text-splitters

# Vector DB
chromadb

# APIs
tavily-python
amadeus

